requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
biopython=1.78=py37he774522_0
blas=1.0=mkl
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2020.12.8=haa95532_0
certifi=2020.12.5=py37haa95532_0
cffi=1.14.4=py37hcd4344a_0
chardet=4.0.0=py37haa95532_1003
cryptography=3.3.1=py37hcd4344a_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py37_0
freetype=2.10.4=hd328e21_0
graphviz=2.38=hfd603c8_2
html5lib=1.1=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
intel-openmp=2020.2=254
isodate=0.6.0=py_1
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
keepalive=0.5=py37hc8dfbb8_4
kiwisolver=1.3.0=py37hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.6.0=h1a1b453_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.2=hf4a77e7_3
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py37hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mysql-connector-c=6.1.11=h33f27b4_1
mysql-connector-python=8.0.18=py37h13ed8b8_1
ninja=1.10.2=py37h6d14046_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
olefile=0.46=py37_0
openssl=1.1.1i=h2bbff1b_0
pandas=1.2.0=py37hf11a4ad_0
pillow=8.1.0=py37h4fa10fc_0
pip=20.3.3=py37haa95532_0
protobuf=3.6.0=py37he025d50_0
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python-graphviz=0.15=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda102_cudnn7_0
pytz=2020.5=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
rdflib=5.0.0=py37hc8dfbb8_3
requests=2.25.1=pyhd3eb1b0_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
setuptools=51.0.0=py37haa95532_2
sip=4.19.8=py37h6538335_0
six=1.15.0=py37haa95532_0
sparqlwrapper=1.8.5=py37hc8dfbb8_1004
sqlite=3.33.0=h2a8f88b_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu102
tornado=6.1=py37h2bbff1b_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
webencodings=0.5.1=py37_1
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
